Removed the backend endpoint for deleting users
Users cannot be deleted from the frontend, and hence, we no longer need an API endpoint for deleting users.
Also added logic to remove the user password property before sending the user object.
Additionally, when the password of a user is incorrect for a login attempt, a 401 is now returned instead of a 400 status.